+++ crates/apps/build.rs
@@ -1,12 +1,10 @@
+use std::env;
 use std::fs::File;
 use std::io::Write;
-use std::str;
+use std::path::PathBuf;
 
 use cargo_metadata::MetadataCommand;
 use git2::{DescribeFormatOptions, DescribeOptions, Repository};
-
-/// Path to the .proto source files, relative to `apps_lib` directory
-const PROTO_SRC: &str = "./proto";
 
 fn main() {
     // Discover the repository version, if it exists
@@ -23,8 +21,10 @@ fn main() {
         Some(describe) => describe.format(Some(&describe_format)).ok(),
         None => None,
     };
+    let out_dir = env::var("OUT_DIR").unwrap();
     let mut version_rs =
-        File::create("./version.rs").expect("cannot write version");
+        File::create(PathBuf::from_iter([&out_dir, "version.rs"]))
+            .expect("cannot write version");
     let pre = "pub fn namada_version() -> &'static str { \"";
     let post = "\" }";
     match version_string {
@@ -47,7 +47,7 @@ fn main() {
                     "--locked".to_string(),
                     "--offline".to_string(),
                 ])
-                .manifest_path("../apps/Cargo.toml")
+                .manifest_path("./Cargo.toml")
                 .exec()
                 .ok()
                 .and_then(|metadata| {
@@ -69,7 +69,4 @@ fn main() {
                 .expect("cannot write version");
         }
     };
-
-    // Tell Cargo that if the given file changes, to rerun this build script.
-    println!("cargo:rerun-if-changed={}", PROTO_SRC);
 }