+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## v1.1.2
+
+Namada apps v1.1.2 patch release.
+
+### MISCELLANEOUS
+
+- Updated Namada libraries to lib-v0.47.2
+  ([\#4363](https://github.com/anoma/namada/pull/4363))
+
 ## v1.1.1
 
 Namada apps v1.1.1 patch release.
@@ -26,6 +35,15 @@ Namada v1.0.0 is the first stable release of the Namada apps.
 
 - updated namada libraries to libs-v0.46.1
   ([\#4123](https://github.com/anoma/namada/pull/4123))
+
+## libs-v0.47.2
+
+Namada libs v0.47.2 patch release.
+
+### IMPROVEMENTS
+
+- Print tx's events in output of `namadac tx-result`.
+  ([\#4347](https://github.com/anoma/namada/pull/4347))
 
 ## libs-v0.47.1